Change miles of trails form int to double 07/06/2016@11:25

+++ src/main/java/com/techelevator/model/JDBCParkDao.java
@@ -41,7 +41,7 @@ public class JDBCParkDao implements ParkDao {
 		park.setState(results.getString("state"));
 		park.setAcreage(results.getInt("acreage"));
 		park.setElevationInFeet(results.getInt("elevationInFeet"));
-		park.setMilesOfTrail(results.getInt("milesOfTrail"));
+		park.setMilesOfTrail(results.getDouble("milesOfTrail"));
 		park.setNumberOfCampsites(results.getInt("numberOfCampsites"));
 		park.setClimate(results.getString("climate"));
 		park.setYearFounded(results.getInt("yearFounded"));
